add images in list
