+++ app/src/main/res/layout/list_item_card.xml
@@ -3,6 +3,8 @@
     android:orientation="vertical" android:layout_width="match_parent"
     android:layout_height="match_parent">
 
+
+
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
